[refactor] Finally got around to seperating User & User Manager class
